feat: added feature(modify/update) for Member

+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -5,6 +5,7 @@ import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -36,14 +37,23 @@ public class MemberController {
     }
 
     @PutMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> updateMember(@PathVariable Long id, @RequestBody MemberRequest param) {
-        memberService.updateMember(id, param);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<MemberResponse> updateMember(
+            @PathVariable Long id,
+            @RequestBody MemberRequest param,
+            @AuthenticationPrincipal UserDetail userDetail) {
+        if (id.equals(userDetail.getId())) {
+            memberService.updateMember(id, param);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @DeleteMapping("/members/{id}")
-    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id) {
-        memberService.deleteMember(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<MemberResponse> deleteMember(@PathVariable Long id, @AuthenticationPrincipal UserDetail userDetail) {
+        if (id.equals(userDetail.getId())) {
+            memberService.deleteMember(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 }
